+++ src/main/java/io/rediset/exception/CommandArityException.java
@@ -1,0 +1,9 @@
+package io.rediset.exception;
+
+/** Raised when a command is invoked with the wrong number of arguments. */
+public class CommandArityException extends CommandException {
+
+    public CommandArityException(String commandName) {
+        super("ERR", "wrong number of arguments for '" + commandName + "' command");
+    }
+}